manual mode will trigger critical mode too
Reason:
The origin code left manual mode to its own. Even after the temperature reaches
the roof, the fan speed won't change which seems not appropriate whether
you have a firefox situation as I do or not.
I want the fan to be quiet when not using firefox, cool down fast after
using firefox and be quiet again. Auto mode is not quiet enough. So I need
a quiet fan speed in manual mode when it's cool, and a critical mode one for
fast cool down.

Changes:
1. Manual mode can trigger critical mode. The boolen "runing" is removed,
because it's always runing now.
2. Add speed 0(turn off the fan).
3. Add "min sane critical speed". Critical speed is no smaller than this(we
have speed 0 now, it's not appropriate for critical speed).